pyproject.toml
[tool.poetry]
name = "accounts"
version = "0.1.0"
description = "Accounts system using OAuth2 and for payment services"
authors = ["begench <begenchallakov070@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
django = "^5.0.1"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.3.1"
ua-parser = "^0.18.0"
user-agents = "^2.2.0"
django-user-agents = "^0.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-oauth-toolkit = "^2.3.0"
django-jet-reboot = "^1.3.7"
psycopg2-binary = "^2.9.9"
emoji = "^2.10.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}

[tool.poetry.group.dev.dependencies]
python = "^3.10"
requests = "^2.31.0"
django = "^5.0.1"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.3.1"
ua-parser = "^0.18.0"
user-agents = "^2.2.0"
django-user-agents = "^0.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-oauth-toolkit = "^2.3.0"
psycopg2-binary = "^2.9.9"
django-jet-reboot = "^1.3.7"
emoji = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
